CSS Home page
Writing down the CSS for the home page, dreading having to make it responsive but the base is there. This project I have set out to make everything within a CSS grid, which is proving to be more of a challenge when wanting a responsive design that adhears to a grid system.